split flutter bridges into common and ui
