+++ android/app/src/main/kotlin/io/rebble/cobble/bridges/common/Connection.kt
@@ -1,8 +1,10 @@
-package io.rebble.cobble.bridges
+package io.rebble.cobble.bridges.common
 
 import io.rebble.cobble.bluetooth.ConnectionLooper
 import io.rebble.cobble.bluetooth.ConnectionState
 import io.rebble.cobble.bluetooth.watchOrNull
+import io.rebble.cobble.bridges.FlutterBridge
+import io.rebble.cobble.bridges.ui.BridgeLifecycleController
 import io.rebble.cobble.data.toPigeon
 import io.rebble.cobble.datasources.WatchMetadataStore
 import io.rebble.cobble.pigeons.BooleanWrapper
