+++ android/app/src/main/kotlin/io/rebble/cobble/bridges/ui/AppLifecycle.kt
@@ -1,7 +1,8 @@
-package io.rebble.cobble.bridges
+package io.rebble.cobble.bridges.ui
 
 import io.flutter.plugin.common.BinaryMessenger
 import io.rebble.cobble.MainActivity
+import io.rebble.cobble.bridges.FlutterBridge
 import io.rebble.cobble.pigeons.BooleanWrapper
 import io.rebble.cobble.pigeons.toMapExt
 import io.rebble.cobble.util.registerAsyncPigeonCallback
