+++ android/app/src/main/kotlin/io/rebble/cobble/bridges/ui/Debug.kt
@@ -1,6 +1,7 @@
-package io.rebble.cobble.bridges
+package io.rebble.cobble.bridges.ui
 
 import android.content.Context
+import io.rebble.cobble.bridges.FlutterBridge
 import io.rebble.cobble.log.collectAndShareLogs
 import io.rebble.cobble.pigeons.Pigeons
 import javax.inject.Inject
