+++ android/app/src/main/kotlin/io/rebble/cobble/di/ActivitySubcomponent.kt
@@ -4,17 +4,26 @@ import dagger.BindsInstance
 import dagger.Subcomponent
 import io.rebble.cobble.MainActivity
 import io.rebble.cobble.bridges.FlutterBridge
+import io.rebble.cobble.di.bridges.CommonBridge
+import io.rebble.cobble.di.bridges.CommonBridgesModule
+import io.rebble.cobble.di.bridges.UiBridge
+import io.rebble.cobble.di.bridges.UiBridgesModule
 import javax.inject.Scope
 
 @PerActivity
 @Subcomponent(
         modules = [
             ActivityModule::class,
-            FlutterBridgesModule::class
+            CommonBridgesModule::class,
+            UiBridgesModule::class,
         ]
 )
 interface ActivitySubcomponent {
-    fun createFlutterBridges(): Set<FlutterBridge>
+    @CommonBridge
+    fun createCommonBridges(): Set<FlutterBridge>
+
+    @UiBridge
+    fun createUiBridges(): Set<FlutterBridge>
 
     @Subcomponent.Factory
     interface Factory {
